Assume nbxmpp availability in all plugins
Trying to import non-existent common.xmpp would trigger an exception were it
not for a a side-effect of demandimport that Gajim uses. It's not nice to rely
on it in gotr and juick.

juick does not use nbxmpp for anything. Let's just drop the import.

whiteboard does a Gajim version check that no other plugin does.

nbxmpp has been required since Gajim 0.16. In all plugins let's just assume
it's available. Some plugins already do that.